questions controller and container update
