+++ app/javascript/react/containers/QuestionShowContainer.js
@@ -4,13 +4,45 @@ import QuestionShowTile from "../components/QuestionShowTile"
 class QuestionShowContainer extends Component {
   constructor(props) {
     super(props);
-    this.state = {}
+    this.state = {
+      question: "",
+      answers: []
+    }
   }
+
+  componentDidMount() {
+      let quiz_id =window.location.href.match(/quizzes\/\d+/);
+      quiz_id = quiz_id[0].match(/\d+/);
+      let  question_id =window.location.href.match(/questions\/\d+/) ;
+      question_id = question_id[0].match(/\d+/);
+
+        fetch('/api/v1/quizzes/'+ quiz_id + '/questions/' + question_id)
+          .then(response => {
+            if (response.ok) {
+              return response
+            } else {
+              let errorMessage = `${response.status} (${response.statusText})`,
+                error = new Error(errorMessage)
+              throw error
+            }
+          })
+          .then(response => response.json())
+          .then(body => {
+            this.setState({
+              question: body.question.body,
+              answers: body.question.answers
+            })
+          })
+          .catch(error => console.error(`Error in fetch: ${error.message}`))
+      }
 
     render() {
       return(
         <div>
-          <QuestionShowTile />
+          <QuestionShowTile
+            question = {this.state.question}
+            answers = {this.state.answers}
+          />
         </div>
       )
     }
